srv/requirements.txt
bottle
pandas
snowflake-connector-python[pandas]
